+++ demo/DemoIsolated/Client/MaintainanceMyClient.generated.cs
@@ -1,0 +1,72 @@
+using Caffoa;
+using Caffoa.Defaults;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading;
+using System.Linq;
+using System.Net.Http;
+using System.Net.Mime;
+using System.Collections.Specialized;
+using System.Threading.Tasks;
+using System.Text;
+using Microsoft.AspNetCore.Http.Extensions;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
+using static System.FormattableString;
+using DemoIsolated.Model.Base;
+using DemoIsolated.Model;
+
+#nullable enable
+
+namespace demo.DemoIsolated.Client
+{
+    /// AUTO GENERATED CLASS
+    public partial class MaintainanceMyClient
+    {
+        /// <summary>
+        /// all servers from the openapi definition  
+        /// </summary>
+        public static string[] Servers => new string[] { "https://api.demoserver.cloud", "https://{prefix}.testserver.cloud/{version}" };
+        
+        private string _baseUri = null!;
+        public string BaseUri {
+            get => _baseUri;
+            set => _baseUri = value.EndsWith("/") ? value : $"{value}/";
+        }
+        public HttpClient Client { get; }
+        public ILogger Logger { get; }
+        public ICaffoaParseErrorHandler ErrorHandler  { get; }
+        public ICaffoaJsonParser JsonParser { get; }
+        public ICaffoaJsonSerializer JsonSerializer { get; }
+        public MaintainanceMyClient(string baseUri, HttpClient? client = null, ILogger? logger = null, ICaffoaParseErrorHandler? errorHandler = null, ICaffoaJsonParser? jsonParser = null, ICaffoaJsonSerializer? jsonSerializer = null) {
+            BaseUri = baseUri;
+            Client = client ?? new HttpClient();
+            Logger = logger ?? NullLogger.Instance;
+            JsonSerializer = jsonSerializer ?? new DefaultCaffoaResultHandler();
+            ErrorHandler = errorHandler ?? new DefaultCaffoaErrorHandler(Logger, JsonSerializer);
+            JsonParser = jsonParser ?? new DefaultCaffoaJsonParser(ErrorHandler);
+        }
+        partial void PrepareRequest(HttpRequestMessage msg);
+        partial void ProcessResponse(HttpResponseMessage msg);
+
+        /// <summary>
+        /// start a long running function via durable functions
+        /// 202 -> started long running function
+        /// </summary>
+        public virtual async Task<IsoLongRunningfunctionStatus> LongRunningFunctionAsync(Guid id, CancellationToken cancellationToken = default) {
+            var uriBuilder = new UriBuilder(Invariant($"{BaseUri}startLongRunningFunction/{id}"));
+            using var httpRequest = new HttpRequestMessage(HttpMethod.Post, uriBuilder.ToString());
+            PrepareRequest(httpRequest);
+            using var httpResult = await Client.SendAsync(httpRequest, cancellationToken);
+            ProcessResponse(httpResult);
+            if(!httpResult.IsSuccessStatusCode) {
+                var errorData = await httpResult.Content.ReadAsStringAsync(cancellationToken);
+                throw new CaffoaWebClientException((int)httpResult.StatusCode, errorData);
+            }
+            await using var resultStream = await httpResult.Content.ReadAsStreamAsync(cancellationToken);
+            var resultObject = await JsonParser.Parse<IsoLongRunningfunctionStatus>(resultStream);
+            return resultObject;
+        }
+    }
+}