+++ demo/DemoV2.AspNetNewtonSoft/Services/DemoV2UserService.cs
@@ -27,6 +27,11 @@ namespace DemoV2.AspNetNewtonSoft.Services
             return Task.CompletedTask;
         }
 
+        public Task UploadImage2Async(string userId, Stream stream, CancellationToken cancellationToken = default)
+        {
+            throw new NotImplementedException();
+        }
+
         public async Task<IEnumerable<ASPNUser>> UsersGetByBirthdateAsync(DateOnly date,
             CancellationToken cancellationToken = default)
         {