+++ demo/DemoV2.Text.Json/Services/DemoV2UserService.cs
@@ -34,6 +34,11 @@ namespace DemoV2.Text.Json.Services
             return Task.CompletedTask;
         }
 
+        public Task UploadImage2Async(string userId, Stream stream, CancellationToken cancellationToken = default)
+        {
+            throw new NotImplementedException();
+        }
+
         public async Task<IEnumerable<STJUser>> UsersGetByBirthdateAsync(DateOnly date,
             CancellationToken cancellationToken = default)
         {